+++ Algoritmo18/src/main/java/br/ufg/inf/es/construcao/algoritmo18/Algoritmo18.java
@@ -3,9 +3,24 @@ package br.ufg.inf.es.construcao.algoritmo18;
 import br.ufg.inf.es.construcao.algoritmo05.Algoritmo05;
 import br.ufg.inf.es.construcao.algoritmo16.Algoritmo16;
 
+/**
+ *
+ * @author Felipe
+ */
 public class Algoritmo18 {
 
-    public static int logNaturalSimples(int n, int k) {
+    /**
+     * Encontra o log de 'k' para a base 'n'.
+     *
+     * @param n É a base.
+     * @param k É o logaritmando.
+     *
+     * @return Retorna o fatorial de 'n'
+     *
+     * @throws IllegalArgumentException Se o parâmetro "n" for menor que 1 ou 
+     * se "k" for menor que 2.
+     */
+    public static double logNaturalSimples(int n, int k) {
 
         if (n < 1) {
             throw new IllegalArgumentException("O valor de 'n' deve ser maior "
@@ -17,7 +32,7 @@ public class Algoritmo18 {
         }
 
         int i = 2;
-        int e = n + 1;
+        double e = n + 1;
 
         while (i <= k) {
